services/cyber_intel/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
dnspython==2.4.2
python-dateutil==2.8.2
aiohttp==3.9.1
requests==2.31.0
python-multipart==0.0.6
